\name{EnsDb.Athaliana.vAtRTD2}
\docType{package}

\alias{EnsDb.Athaliana.vAtRTD2-package}
\alias{EnsDb.Athaliana.vAtRTD2}
\alias{EnsDb.Athaliana.vAtRTD2}


\title{Ensembl based annotation package}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at Ensembl on
  Tue Apr 30 19:03:01 2019 and based on the AtRTD2
}

\author{Wenbin Guo}



\examples{
## load the library
##library(EnsDb.Athaliana.vAtRTD2)
## list the contents that are loaded into memory
ls('package:EnsDb.Athaliana.vAtRTD2')
## show the db object that is loaded by calling it's name
EnsDb.Athaliana.vAtRTD2

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
